test(bft): move global-GUARD test to integration binary
Putting full_guard_lifecycle_through_global in the same #[cfg(test)] mod
as the pure-helper tests poisoned messages::tests::test_proposal_sign_verify
in CI: the OnceLock GUARD remained initialised after my test ran, and the
proposal/prevote/precommit sign() calls in messages tests then hit the live
guard with their own (h, r) tuples. Two failures, race-order dependent.

Integration tests get their own test binary, so the OnceLock state is
isolated. Same coverage, no contamination.
